changes in vectorColorsBranch ...WORKS NOW!
Made 4 changes to get the vector coloring working (meaning the vectors are plotted red if the particle has a flow path that collides with the filter)
a): in %% Check if individual particle displacement vectors penetrate filter wall
if dip_array(stackFilt(point(1),point(2),point(3)))==1 instead of if(stackFilt(point)==1)
b) added break command in that if-condition
c) in %% Display fixed the elseif any (x + y + z < 0)  condition so that the vectors are display properly

d) Has nothng to do with the rest. A major change that should be in master branch too.
In %% Calculation replaced found = (sortrows(found',1))'; with if-condition tht checks if found is empty (found=[ ]). If so the algorithm would crash without that if-condition. found may be empty if dummy images are used for testing the algorithm.